separated details window into its own file
